graphs/requirements.txt
pyserial==3.5
matplotlib==3.8.2
numpy==1.26.2
